Break-up package into meaningful modules
